+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 12.88.0
+
+* `[pager]` Added new package for pager (`less`/`more`) setup
+* `[terminal/tty]` Added new package for working with TTY
+* `[fmtc]` Added method `IsColorsSupported`
+
 ### 12.87.0
 
 * `[fmtc]` Added tag for italic text (`{&}`)